# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для аренды квартир",
        "title": "Renting App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000/api",
    "basePath": "/",
    "paths": {
...
        "/apartment/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить список всех квартир пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apartments"
                ],
                "summary": "Получить все квартиры",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать новую квартиру",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apartments"
                ],
                "summary": "Создать квартиру",
                "parameters": [
                    {
                        "description": "Apartment input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateApartmentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apartment/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить квартиру по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apartments"
                ],
                "summary": "Получить квартиру по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Apartment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetApartmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удалить квартиру по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apartments"
                ],
                "summary": "Удалить квартиру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Apartment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновить данные по квартире",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apartments"
                ],
                "summary": "Обновить квартиру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Apartment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Apartment update input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateApartmentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateApartmentInput": {
            "type": "object",
            "required": [
                "city",
                "construction_year",
                "floor",
                "latitude",
                "longitude",
                "rooms",
                "street"
            ],
            "properties": {
                "area": {
                    "type": "integer"
                },
                "bathroom_type": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "concierge": {
                    "type": "boolean"
                },
                "construction_type": {
                    "type": "string"
                },
                "construction_year": {
                    "type": "integer"
                },
                "elevator": {
                    "type": "boolean"
                },
                "floor": {
                    "type": "integer"
                },
                "garbage_chute": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "remont": {
                    "type": "string"
                },
                "rooms": {
                    "type": "integer"
                },
                "street": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetApartmentPhotoResponse": {
            "type": "object",
            "properties": {
                "apartment_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_cover": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.GetApartmentResponse": {
            "type": "object",
            "properties": {
                "apartment_number": {
                    "type": "string"
                },
                "apartment_photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetApartmentPhotoResponse"
                    }
                },
                "area": {
                    "type": "integer"
                },
                "bathroom_type": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "concierge": {
                    "type": "boolean"
                },
                "construction_type": {
                    "type": "string"
                },
                "construction_year": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "elevator": {
                    "type": "boolean"
                },
                "floor": {
                    "type": "integer"
                },
                "garbage_chute": {
                    "type": "boolean"
                },
                "house": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "remont": {
                    "type": "string"
                },
                "rooms": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateApartmentInput": {
            "type": "object",
            "properties": {
                "apartment_number": {
                    "type": "string"
                },
                "area": {
                    "type": "integer"
                },
                "bathroom_type": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "concierge": {
                    "type": "boolean"
                },
                "construction_type": {
                    "type": "string"
                },
                "construction_year": {
                    "type": "integer"
                },
                "district": {
                    "type": "string"
                },
                "elevator": {
                    "type": "boolean"
                },
                "floor": {
                    "type": "integer"
                },
                "garbage_chute": {
                    "type": "boolean"
                },
                "house": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "remont": {
                    "type": "string"
                },
                "rooms": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid input data"
                }
            }
        },
...
    }
}